* upgrade Mustache version to 0.9.10 * adding SafeMustacheCustomVisitorFactory * update others to use it also
