+++ pom.xml
@@ -143,7 +143,7 @@
             <dependency>
                 <groupId>com.github.spullara.mustache.java</groupId>
                 <artifactId>compiler</artifactId>
-                <version>0.9.2</version>
+                <version>0.9.10</version>
             </dependency>
   			<dependency>
                 <groupId>org.apache.commons</groupId>
