+++ sawmill-core/src/main/java/io/logz/sawmill/mustache/factories/UnescapedMustacheFactory.java
@@ -1,10 +1,10 @@
-package io.logz.sawmill;
+package io.logz.sawmill.mustache.factories;
 
 import com.github.mustachejava.DefaultMustacheFactory;
 import com.github.mustachejava.MustacheException;
 import com.github.mustachejava.reflect.ReflectionObjectHandler;
-import io.logz.sawmill.utilities.JsonUtils;
 
+import com.google.common.collect.ImmutableSet;
 import java.io.IOException;
 import java.io.Writer;
 import java.util.List;
@@ -12,9 +12,9 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class UnescapedWithJsonStringMustacheFactory extends DefaultMustacheFactory {
-    public UnescapedWithJsonStringMustacheFactory() {
-        super();
+public class UnescapedMustacheFactory extends SafeMustacheCustomVisitorFactory {
+    public UnescapedMustacheFactory() {
+        super(ImmutableSet.of(), ".");
 
         this.setObjectHandler(new ListTransformObjectHandler());
     }
@@ -29,15 +29,6 @@ public class UnescapedWithJsonStringMustacheFactory extends DefaultMustacheFacto
     }
 
     public class ListTransformObjectHandler extends ReflectionObjectHandler {
-        @Override
-        public String stringify(Object object) {
-            if (object instanceof Map) {
-                return JsonUtils.toJsonString(object);
-            }
-
-            return super.stringify(object);
-        }
-
         @Override
         public Object coerce(final Object object) {
             if (object != null && object instanceof List) {
